Fixed noteDescription edit text line bug

+++ app/src/main/java/com/neuralbit/letsnote/ArchivedNote.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "ArchivedNotesTable")
 class ArchivedNote(
-    @PrimaryKey @ColumnInfo(name = "id")val id : Int
+    @PrimaryKey @ColumnInfo(name = "id") val id: Long
 )
