+++ app/src/main/java/com/neuralbit/letsnote/PinnedNote.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "PinnedNotesTable")
 class PinnedNote(
-    @PrimaryKey @ColumnInfo(name = "id")val id : Int
+    @PrimaryKey @ColumnInfo(name = "id") val id: Long
 )
